+++ src/lib/terrain/heightmap/heightmap-cpu/heightmap-node-geometry.ts
