+++ src/test/test-terrain.ts
@@ -7,7 +7,7 @@ import {
     EBoardSquareType,
     EComputationMethod,
     EVoxelsDisplayMode,
-    HeightmapViewer,
+    HeightmapViewerCpu,
     InstancedBillboard,
     PromisesQueue,
     TerrainViewer,
@@ -80,7 +80,7 @@ class TestTerrain extends TestTerrainBase {
             });
         }, 150);
 
-        const heightmapViewer = new HeightmapViewer(map, {
+        const heightmapViewer = new HeightmapViewerCpu(map, {
             basePatchSize: chunkSize.xz,
             maxLevel: 5,
             voxelRatio: 2,
